마스터 새 함수 인자 조정 int -> float

+++ DVDnBluray/src/DnB/DTO/TestDTO.java
@@ -5,8 +5,8 @@ package DnB.DTO;
 // Git Test Class
 public class TestDTO {
 
-	public void myfun(int param) {
-		
+	public void myfun(float param) {
+		// 인수 타입 조정 int -> float
 	}
 	
 }
